pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "parsec"
version = "0.1.0"
description = "Natural language cloud cost investigation tool"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.35.0",
    "pydantic>=2.8.0",
    "dynaconf>=3.2.0",
    "pyyaml>=6.0",
    "anthropic[vertex,bedrock]>=0.40.0",
    "asyncpg>=0.31.0",
    "boto3>=1.34.0",
    "azure-identity>=1.15.0",
    "azure-storage-blob>=12.19.0",
    "google-cloud-bigquery>=3.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "pytest-mock>=3.12",
    "httpx>=0.27",
    "black>=23.12",
    "ruff>=0.1.9",
    "mypy>=1.8",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ["py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "B", "SIM", "I", "N", "C90"]
ignore = ["E501", "B008", "B904"]

[tool.ruff.lint.mccabe]
max-complexity = 12

[tool.mypy]
python_version = "3.11"
strict_optional = true
no_implicit_reexport = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "boto3.*",
    "botocore.*",
    "google.*",
    "azure.*",
    "asyncpg.*",
    "anthropic.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]
